40-m1-implement-a-folder-structure: moved functions to modules
